requirements.txt
# Core Streamlit and visualization
streamlit>=1.32.0
plotly>=5.19.0
pandas>=2.1.0
numpy>=1.26.0

# LLM API clients
openai>=1.40.0
anthropic>=0.34.0
google-generativeai>=0.7.0

# HTTP and JSON handling
httpx>=0.25.0
requests>=2.31.0
jsonschema>=4.20.0

# Sentiment analysis
vaderSentiment>=3.3.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0